+++ extensions/twenty/src/components/FieldComponent.tsx
@@ -3,6 +3,7 @@ import { DataModelField } from "../services/zod/schema/recordFieldSchema";
 import TextInput from "./TextInput";
 import Select from "./Select";
 import Rating from "./Rating";
+import MultiSelect from "./MultiSelect";
 
 type FieldComponentProps = {
   values: {
@@ -25,6 +26,12 @@ export default function FieldComponent({ values }: FieldComponentProps) {
     case "RATING": {
       return <Rating values={{ field }} {...itemProps} />;
     }
+    case "MULTI_SELECT": {
+      return <MultiSelect values={{ field }} {...itemProps} />;
+    }
+    // case "MULTI_SELECT": {
+    //   return <MultiSelect values={{ field }} />;
+    // }
     default:
       return <></>;
   }