+++ extensions/twenty/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Twenty Changelog
 
+## [Enhancements] - {PR_MERGE_DATE}
+
+- Added support for multi-select field type
+
 ## [Enhancements] - 2024-12-18
 
 - Updated select field type with preference